requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-bootstrap-datepicker-plus==4.0.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
whitenoise==6.2.0
